pyproject.toml
[project]
name = "langchain-agentic-rag-sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.8",
    "langchain-openai>=1.0.3",
    "numpy>=2.3.5",
    "python-dotenv>=1.2.1",
]
